Moves Carousel class to models module.
